+++ android/src/main/java/com/pspdfkit/flutter/pspdfkit/ConfigurationAdapter.java
@@ -1,5 +1,5 @@
 /*
- *   Copyright © 2018-2022 PSPDFKit GmbH. All rights reserved.
+ *   Copyright © 2018-2023 PSPDFKit GmbH. All rights reserved.
  *
  *   THIS SOURCE CODE AND ANY ACCOMPANYING DOCUMENTATION ARE PROTECTED BY INTERNATIONAL COPYRIGHT LAW
  *   AND MAY NOT BE RESOLD OR REDISTRIBUTED. USAGE IS BOUND TO THE PSPDFKIT LICENSE AGREEMENT.
@@ -29,6 +29,7 @@ import com.pspdfkit.configuration.page.PageScrollMode;
 import com.pspdfkit.configuration.settings.SettingsMenuItemType;
 import com.pspdfkit.configuration.sharing.ShareFeatures;
 import com.pspdfkit.configuration.theming.ThemeMode;
+import com.pspdfkit.preferences.PSPDFKitPreferences;
 import com.pspdfkit.ui.special_mode.controller.AnnotationTool;
 
 import java.util.ArrayList;
@@ -180,6 +181,12 @@ class ConfigurationAdapter {
 
     // Instant Options
     private static final String ENABLE_INSTANT_COMMENTS = "enableInstantComments";
+
+    // Measurement tools options
+    private static final String ENABLED_MEASUREMENT_TOOLS = "enableMeasurementTools";
+    private static final String ENABLE_MAGNIFIER = "enableMagnifier";
+    private static final String ENABLED_MEASUREMENT_TOOL_SNAPPING = "enableMeasurementToolSnapping";
+
 
     @NonNull
     private final PdfActivityConfiguration.Builder configuration;
@@ -363,8 +370,24 @@ class ConfigurationAdapter {
             if (key != null) {
                 enableInstantComments = (boolean) configurationMap.get(key);
             }
+
+            key = getKeyOfType(configurationMap, ENABLE_MAGNIFIER, Boolean.class);
+            if (key != null) {
+                configureMagnifierEnabled((Boolean) configurationMap.get(key));
+            }
+
+            key = getKeyOfType(configurationMap, ENABLED_MEASUREMENT_TOOLS, Boolean.class);
+            if (key != null) {
+                configureMeasurementToolsEnabled((Boolean) configurationMap.get(key));
+            }
+
+            key = getKeyOfType(configurationMap, ENABLED_MEASUREMENT_TOOL_SNAPPING, Boolean.class);
+            if (key != null) {
+                configureMeasurementToolSnappingEnabled(context,(Boolean) configurationMap.get(key));
+            }
         }
     }
+
 
     private void configurePageTransition(@NonNull final String transition) {
         switch (transition) {
@@ -709,6 +732,18 @@ class ConfigurationAdapter {
 
     private void configureAutosaveEnabled(boolean autosaveEnabled) {
         configuration.autosaveEnabled(autosaveEnabled);
+    }
+
+    private void configureMeasurementToolsEnabled(Boolean aBoolean) {
+        configuration.setMeasurementToolsEnabled(aBoolean);
+    }
+
+    private void configureMagnifierEnabled(Boolean aBoolean) {
+        configuration.enableMagnifier(aBoolean);
+    }
+
+    private void configureMeasurementToolSnappingEnabled(Context context,Boolean aBoolean) {
+        PSPDFKitPreferences.get(context).setMeasurementSnappingEnabled(aBoolean);
     }
 
     private <T> boolean containsKeyOfType(@NonNull HashMap<String, Object> configurationMap,