+++ ios/Classes/PspdfkitPlugin.m
@@ -1,5 +1,5 @@
 //
-//  Copyright © 2018-2022 PSPDFKit GmbH. All rights reserved.
+//  Copyright © 2018-2023 PSPDFKit GmbH. All rights reserved.
 //
 //  THIS SOURCE CODE AND ANY ACCOMPANYING DOCUMENTATION ARE PROTECTED BY INTERNATIONAL COPYRIGHT LAW
 //  AND MAY NOT BE RESOLD OR REDISTRIBUTED. USAGE IS BOUND TO THE PSPDFKIT LICENSE AGREEMENT.
@@ -45,8 +45,11 @@ PSPDFSettingKey const PSPDFSettingKeyHybridEnvironment = @"com.pspdfkit.hybrid-e
         NSString *iOSLicenseKey = call.arguments[@"iOSLicenseKey"];
         [PSPDFKitGlobal setLicenseKey:iOSLicenseKey options:@{PSPDFSettingKeyHybridEnvironment: @"Flutter"}];
     }else if ([@"present" isEqualToString:call.method]) {
+        
         NSString *documentPath = call.arguments[@"document"];
         
+      
+      
         if (documentPath == nil || documentPath.length <= 0) {
             FlutterError *error = [FlutterError errorWithCode:@"" message:@"Document path may not be nil or empty." details:nil];
             result(error);
@@ -61,7 +64,21 @@ PSPDFSettingKey const PSPDFSettingKeyHybridEnvironment = @"com.pspdfkit.hybrid-e
             result(error);
             return;
         }
+        NSDictionary *measurementScale = call.arguments[@"measurementScale"];
+        if (measurementScale){
+            PSPDFMeasurementScale *scale = [PspdfkitMeasurementConvertor convertScaleWithMeasurement:measurementScale];
+            if (scale != nil) {
+                document.measurementScale = scale;
+            }
+        }
         
+        NSString *measurementPrecision = call.arguments[@"measurementPrecision"];
+        
+        if (measurementPrecision){
+            PSPDFMeasurementPrecision precision = [PspdfkitMeasurementConvertor convertPrecisionWithPrecision:measurementPrecision];
+            document.measurementPrecision = precision;
+        }
+       
         [PspdfkitFlutterHelper unlockWithPasswordIfNeeded:document dictionary:configurationDictionary];
         
         BOOL isImageDocument = [PspdfkitFlutterHelper isImageDocument:documentPath];