+++ lib/configuration_options.dart
@@ -1,5 +1,5 @@
 ///
-///  Copyright © 2018-2022 PSPDFKit GmbH. All rights reserved.
+///  Copyright © 2018-2023 PSPDFKit GmbH. All rights reserved.
 ///
 ///  THIS SOURCE CODE AND ANY ACCOMPANYING DOCUMENTATION ARE PROTECTED BY INTERNATIONAL COPYRIGHT LAW
 ///  AND MAY NOT BE RESOLD OR REDISTRIBUTED. USAGE IS BOUND TO THE PSPDFKIT LICENSE AGREEMENT.
@@ -183,3 +183,11 @@ const String pdfGenerationConfigurationHeight = 'height';
 const String pdfGenerationConfigurationOverride = 'override';
 
 const String pdfGenerationConfigurationImagePath = 'imagePath';
+
+const String measurementUnit = 'measurementUnit';
+
+const String enableMeasurementMagnifier = 'enableMagnifier';
+
+const String enableMeasurementTools = 'enableMeasurementTools';
+
+const String enableMeasurementToolSnapping = 'enableMeasurementToolSnapping';