+++ lib/src/measurements/measurement_precision.dart
@@ -1,0 +1,3 @@
+part of pspdfkit;
+
+enum MeasurementPrecision { oneDP, twoDP, threeDP, fourDP, whole }