+++ lib/widgets/pspdfkit_widget_controller.dart
@@ -1,5 +1,5 @@
 ///
-///  Copyright © 2018-2022 PSPDFKit GmbH. All rights reserved.
+///  Copyright © 2018-2023 PSPDFKit GmbH. All rights reserved.
 ///
 ///  THIS SOURCE CODE AND ANY ACCOMPANYING DOCUMENTATION ARE PROTECTED BY INTERNATIONAL COPYRIGHT LAW
 ///  AND MAY NOT BE RESOLD OR REDISTRIBUTED. USAGE IS BOUND TO THE PSPDFKIT LICENSE AGREEMENT.
@@ -9,6 +9,7 @@
 
 import 'dart:async';
 import 'package:flutter/services.dart';
+import 'package:pspdfkit_flutter/pspdfkit.dart';
 
 class PspdfkitWidgetController {
   final MethodChannel _channel;
@@ -80,4 +81,24 @@ class PspdfkitWidgetController {
   /// Saves the document back to its original location if it has been changed.
   /// If there were no changes to the document, the document file will not be modified.
   Future<bool?> save() async => _channel.invokeMethod('save');
+
+  /// Sets the measurement scale of the document.
+  /// The scale is used to convert between real world measurements and points.
+  /// The default scale is 1 inch = 1 inch.
+  /// @param scale The scale to be used for the document.
+  /// @return True if the scale was set successfully, false otherwise.
+  Future<bool?> setMeasurementScale(MeasurementScale scale) async =>
+      _channel.invokeMethod('setMeasurementScale', <String, dynamic>{
+        'measurementScale': scale.toMap(),
+      });
+
+  /// Sets the measurement precision of the document.
+  /// The precision is used to round the measurement values.
+  /// The default precision is 2 decimal places.
+  /// @param precision The precision to be used for the document.
+  /// @return True if the precision was set successfully, false otherwise.
+  Future<bool?> setMeasurementPrecision(MeasurementPrecision precision) async =>
+      _channel.invokeMethod('setMeasurementPrecision', <String, dynamic>{
+        'measurementPrecision': precision.name,
+      });
 }